requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
controlnet-aux>=0.0.7
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
gradio>=4.20.0
huggingface-hub>=0.21.0
safetensors>=0.4.2
scipy>=1.12.0
